COM-1467 - create modal component
