+++ App.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { StatusBar, View, StyleSheet, AppState, Linking, Platform } from 'react-native';
+import { StatusBar, View, StyleSheet, AppState, Text, Platform, Linking } from 'react-native';
 import { createStore } from 'redux';
 import { Provider as ReduxProvider } from 'react-redux';
 import { AppLoading } from 'expo';
@@ -7,11 +7,14 @@ import * as Font from 'expo-font';
 import { Provider as AuthProvider } from './src/context/AuthContext';
 import getEnvVars from './environment';
 import Version from './src/api/version';
-import ConfirmModal from './src/components/modal/ConfirmModal';
 import AppContainer from './src/AppContainer';
 import { WHITE } from './src/styles/colors';
 import reducers from './src/store/index';
 import tron from './src/ReactotronConfig';
+import NiModal from './src/components/modal';
+import NiButton from './src/components/form/Button';
+import { MARGIN } from './src/styles/metrics';
+import { FIRA_SANS_BOLD, FIRA_SANS_REGULAR } from './src/styles/fonts';
 
 const store = createStore(reducers, tron.createEnhancer());
 
@@ -53,14 +56,13 @@ const App = () => {
 
   return (
     <>
-      <ConfirmModal
-        visible={modalOpened}
-        title="Nouvelle version de l'app disponible !"
-        contentText="Merci de mettre votre application à jour pour pouvoir continuer d'utiliser l'application :)"
-        buttonCaption="Mettre à jour"
-        onPress={() => { Linking.openURL(appUrl); }}
-        onRequestClose={() => setModalOpened(false)}
-      />
+      <NiModal visible={modalOpened}>
+        <Text style={styles.title}>Nouvelle version de l'app disponible !</Text>
+        <Text style={styles.contentText}>
+            Merci de mettre votre application à jour pour pouvoir continuer d'utiliser l'application :)
+        </Text>
+        <NiButton style={styles.button} caption="Mettre à jour" onPress={() => { Linking.openURL(appUrl); }} />
+      </NiModal>
       <AuthProvider>
         <ReduxProvider store={store}>
           <View style={styles.statusBar}>
@@ -84,5 +86,20 @@ const styles = StyleSheet.create({
   statusBar: {
     backgroundColor: WHITE,
     height: STATUSBAR_HEIGHT,
+  },
+  title: {
+    ...FIRA_SANS_BOLD.LG,
+    marginBottom: MARGIN.LG,
+    textAlign: 'center',
+  },
+  contentText: {
+    ...FIRA_SANS_REGULAR.MD,
+    textAlign: 'center',
+    marginBottom: MARGIN.LG,
+  },
+  button: {
+    display: 'flex',
+    flexDirection: 'row',
+    justifyContent: 'center',
   },
 });
