+++ src/components/activities/ExitActivityModal.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { View, Text, Modal, StyleSheet, TouchableOpacity } from 'react-native';
+import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
+import NiModal from '../modal';
 import { WHITE, MODAL_BACKDROP_GREY, PINK } from '../../styles/colors';
 import { BORDER_RADIUS, PADDING, MARGIN } from '../../styles/metrics';
 import { FIRA_SANS_BOLD, FIRA_SANS_REGULAR } from '../../styles/fonts';
@@ -15,22 +16,18 @@ const ExitActivityModal = ({
   onPressCancelButton,
   onPressConfirmButton,
 }: ExitActivityModalProps) => (
-  <Modal visible={visible} transparent={true}>
-    <View style={styles.modalContainer}>
-      <View style={styles.modalContent} >
-        <Text style={styles.title}>Es-tu sûr de cela ?</Text>
-        <Text style={styles.contentText}>Tous tes progrès dans la leçon seront perdus.</Text>
-        <View style={styles.buttons}>
-          <TouchableOpacity style={styles.cancelButton} onPress={onPressCancelButton}>
-            <Text style={styles.buttonText}>Annuler</Text>
-          </TouchableOpacity>
-          <TouchableOpacity style={styles.closeButton} onPress={onPressConfirmButton}>
-            <Text style={styles.buttonText}>Quitter</Text>
-          </TouchableOpacity>
-        </View>
-      </View>
+  <NiModal visible={visible}>
+    <Text style={styles.title}>Es-tu sûr de cela ?</Text>
+    <Text style={styles.contentText}>Tous tes progrès dans la leçon seront perdus.</Text>
+    <View style={styles.buttons}>
+      <TouchableOpacity style={styles.cancelButton} onPress={onPressCancelButton}>
+        <Text style={styles.buttonText}>Annuler</Text>
+      </TouchableOpacity>
+      <TouchableOpacity style={styles.closeButton} onPress={onPressConfirmButton}>
+        <Text style={styles.buttonText}>Quitter</Text>
+      </TouchableOpacity>
     </View>
-  </Modal>
+  </NiModal>
 );
 
 const styles = StyleSheet.create({
